% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statRes_class.R
\name{plot_volcano}
\alias{plot_volcano}
\title{Volcano plot function for `statRes` objects}
\usage{
plot_volcano(
  x,
  plot_method = NULL,
  fc_colors = c("red", "black", "green"),
  interactive = FALSE,
  bw_theme = FALSE,
  fc_threshold = NULL,
  custom_theme = NULL
)
}
\arguments{
\item{x}{`statRes` object to be plotted, usually the result of `imd_anova`}

\item{plot_method}{defines which test to use for volcano plots, options are "anova" and "gtest"; required for "combined" statistics method, defaults to method used for testing for "anova" or "gtest" statistics}

\item{fc_colors}{vector of length three with character color values interpretable by ggplot. i.e. c("orange","black", "blue") with the values being used to color negative, non-significant, and positive fold changes respectively}

\item{interactive}{TRUE/FALSE for whether to create an interactive plot using plotly}

\item{bw_theme}{TRUE/FALSE for whether to apply a black-white background theme. Applies only when interactive = FALSE.}

\item{fc_threshold}{optional threshold value for fold change estimates that's added to the volcano plot}

\item{custom_theme}{optional custom theme object recognized by ggplot}
}
\description{
Produces volcano plots that summarize the results contained in a `statRes` object.
}
\examples{
dontrun{
library(pmartR)
library(pmartRdata)
#Transform the data

#Group the data by condition
myproData <- group_designation(omicsData = pro_object, main_effects = c("Condition"))

#Apply the IMD ANOVA filter
imdanova_Filt <- imdanova_filter(omicsData = myproData)
myproData <- applyFilt(filter_object = imdanova_Filt, omicsData = myproData, min_nonmiss_anova=2)

#Implement the IMD ANOVA method and compuate all pairwise comparisons (i.e. leave the `comparisons` argument NULL)
anova_res <- imd_anova(omicsData = myproData, test_method = 'anova')
plot_volcano(anova_res)

imd_res <- imd_anova(omicsData = myproData, test_method = 'gtest')
plot_volcano(imd_res)

imd_anova_res <- imd_anova(omicsData = myproData, test_method = 'comb', pval_adjust='bon')
plot_volcano(imd_anova_res,  plot_method = "gtest", bw_theme = TRUE)
plot_volcano(imd_anova_res, plot_method = "anova", bw_theme = TRUE, interactive = TRUE)

}

}
